Create ranges_calculations.py
This API defines a base AxisRange class that represents a range of values on an axis. It has two attributes, starts and stops, that define the start and end points of the range. It also has a transform attribute that is used to apply a non-linear transformation to the axis values.